backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.19
Pillow>=11.0.0
numpy>=2.2.0
psycopg[binary]>=3.3.0
sqlalchemy>=2.0.36
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
debugpy>=1.8.0
rembg>=2.0.0
opencv-python>=4.8.0
onnxruntime>=1.23.0
